Reconparams changed to single precision; pow() -> powf()

+++ src/MBIRModularUtils.c
@@ -536,9 +536,9 @@ int ReadReconParams(
 	}
 
 	/* calculate derived parameters */
-	reconparams->pow_sigmaX_p = pow(reconparams->SigmaX,reconparams->p);
-	reconparams->pow_sigmaX_q = pow(reconparams->SigmaX,reconparams->q);
-	reconparams->pow_T_qmp    = pow(reconparams->T,reconparams->q - reconparams->p);
+	reconparams->pow_sigmaX_p = powf(reconparams->SigmaX,reconparams->p);
+	reconparams->pow_sigmaX_q = powf(reconparams->SigmaX,reconparams->q);
+	reconparams->pow_T_qmp    = powf(reconparams->T,reconparams->q - reconparams->p);
 	reconparams->SigmaXsq = reconparams->SigmaX * reconparams->SigmaX;
 
 	return(0);
